requirements.txt
django==4
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
pydantic